+++ CHANGELOG.md
@@ -1,5 +1,8 @@
+# v1.2.0 (Oct 17, 2021)
+- The app now remembers your settings
+
 # v1.1.0 (September 9, 2021)
--Fixed ports that were non-numeric
+- Improve logic for retrieving processes by port
 
 # v1.0.1 (August 17, 2021)
 - Fix handling of invalid port numbers for Unix-based systems